requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
colorama==0.4.1
coverage==4.5.4
Django==2.2.4
django-extensions==2.2.1
django-markdown==0.8.4
django-markdownx==2.0.28
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
django-taggit==1.1.0
djangorestframework==3.10.2
importlib-metadata==0.19
Markdown==3.1.1
more-itertools==7.2.0
packaging==19.1
phonenumbers==8.10.16
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.0.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
wcwidth==0.1.7
Werkzeug==0.15.5
wincertstore==0.2
zipp==0.5.2
